Add files via upload
The code uses an API key from OpenAI and web scraping to summarize information regarding a user provided stock ticker symbol and uploading it to a .txt file in the same folder as the .py file and the .env file containing the sensitive API Key. With an API Key in the same directory and token billing enabled, the results can be seen with this code.